fix: resolve compile errors across main, src, tests, and benchmarks projects
Root causes fixed:
- Missing using directives for Core/Domain/Repository/Services/Microsoft.Extensions.* namespaces in src/ tree
- Namespace-containment shadowing: types declared directly in GpuImageProcessing.Core (src) silently
  shadowed GpuImageProcessing.Core.Constants enums (FilterType, ProcessingStatus, TransformType, ImageFormat)
  inside any GpuImageProcessing.Core.* namespace; resolved via explicit using-aliases at namespace scope
- Duplicate type definitions across root/src trees (Constants, PerformanceMonitoringService) and within
  the root tree itself (ProcessingEvent/DeviceStatusChangedEvent, IProcessingMiddleware/MiddlewareContext/
  MiddlewareResult, ValidationResult) resolved via targeted renames, preserving all original members
- ComputeShaderPass now constructed via its constructor instead of an invalid object initializer
- ProcessingResult converted to a record to support "with" expressions; added Status/StartTime/
  CompletionTime/DeviceUsed/ProfileUsed and alias properties expected by callers
- Added missing convenience members (ApplyFilterAsync/ApplyTransformAsync/GetProcessingResultsAsync,
  GetJobProgressAsync, ListFiltersAsync) instead of stripping the API surface that referenced them
- Fixed Silk.NET.OpenCL API misuse (Platform/Device handle types, DeviceInfo enum member names,
  ambiguous overloads) and corrected several typos (FullPath -> FullName, ElapsedSeconds -> Elapsed.TotalSeconds)
- Rescoped the two root-level csproj files' Compile items and fixed an invalid Cloo package version
- Fixed FluentAssertions async-throw assertions in tests to wrap calls via Invoking(...)

Pushed: dotnet build gpu-image-processing.sln now completes with 0 errors.

+++ Core/Services/DeviceService.cs
@@ -8,9 +8,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GpuImageProcessing.Core;
 using GpuImageProcessing.Core.Exceptions;
 using GpuImageProcessing.Core.Models;
-using GpuImageProcessing.Core.Enums;
 using GpuImageProcessing.Services;
 using Microsoft.Extensions.Logging;
 
@@ -109,7 +109,7 @@ namespace GpuImageProcessing.Core.Services
                 OpenCLVersion = gpuDevice.Version,
                 DriverVersion = gpuDevice.Driver,
                 IsAvailable = gpuDevice.IsAvailable,
-                ClockFrequencyMHz = gpuDevice.MaxClockFrequencyMhz,
+                ClockFrequencyMHz = (float)gpuDevice.MaxClockFrequencyMhz,
                 SupportsDoublePrecision = gpuDevice.SupportsDoublePrecision
             };
         }
